Improve our man page's rendering in HTML, PDF, and text
Replace blank lines with comment lines to fix vertical whitespace issues in PDF and text. Move all manx invocation formats to the top of the Synopsis section, because it's visually a lot cleaner, especially in HTML where macro-based indentation doesn't carry over. Indent the bullet points manually with spaces, a little farther than I normally would so it works in the plain text rendering. Use a hack to insert non-breaking spaces before bullet points in HTML, because I don't see any way to indent those lines with CSS, given the page structure mandoc generates.

Our bullet points aren't ideal in the PDF rendering, but including proper bullet-point characters (like from Unicode) in the PDF turns out to be non-trivial, and rendering a '+' on top of an 'o' like mandoc does turns out to be about as good as it gets without Unicode characters, so... oh well.